新增指定up充电功能

+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Interfaces/IDailyTaskApi.cs
@@ -115,13 +115,13 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Interfaces
         /// <summary>
         /// 充电
         /// </summary>
-        /// <param name="elec_num"></param>
-        /// <param name="up_mid"></param>
-        /// <param name="oid"></param>
+        /// <param name="elec_num">充电电池数量（B币*10）,必须在20-99990之间</param>
+        /// <param name="up_mid">充电对象用户UID</param>
+        /// <param name="oid">充电来源代码(空间充电：充电对象用户UID;视频充电：稿件avID)</param>
         /// <param name="csrf"></param>
         /// <returns></returns>
         [Post("/x/ugcpay/trade/elec/pay/quick?elec_num={elec_num}&up_mid={up_mid}&otype=up&oid={oid}&csrf={csrf}")]
-        Task<BiliApiResponse<ChargeResponse>> Charge(decimal elec_num, string up_mid, string oid, string csrf);
+        Task<BiliApiResponse<ChargeResponse>> Charge(int elec_num, string up_mid, string oid, string csrf);
 
         /// <summary>
         /// 充电后留言
